+++ composeApp/src/commonMain/kotlin/com/apps/adrcotfas/goodtime/data/local/LabelDao.kt
@@ -50,7 +50,7 @@ interface LabelDao {
     )
     suspend fun updateLabel(
         newName: String,
-        newColorIndex: Long,
+        newColorIndex: Int,
         newUseDefaultTimeProfile: Boolean,
         newTimerProfileName: String?,
         newIsCountdown: Boolean,
