+++ composeApp/src/commonMain/kotlin/com/apps/adrcotfas/goodtime/stats/TimelineTab.kt
@@ -152,7 +152,7 @@ fun TimelineListItem(
     modifier: Modifier = Modifier,
     session: Session,
     isSelected: Boolean = false,
-    colorIndex: Long,
+    colorIndex: Int,
     onClick: () -> Unit,
     onLongClick: () -> Unit,
 ) {
